requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-firestore==2.11.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
